% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-row.R
\name{blank_row}
\alias{blank_row}
\title{blank_row}
\usage{
blank_row(df)
}
\arguments{
\item{df}{the data frame}
}
\description{
add a row of NA to a data frame.  This is a wrapper around the \code{\link{summary_row}}
function.
}
\examples{
df <- data.frame("year" = 2014:2016,
                 "a" = c(NA, 1, 3),
                 "b" = c("g", "e", "f")
      )

averages_row(df = df, cols = 2)
averages_row(df = df, cols = 2, label_col = 1, label = "Totals")

df$d <- 5:7

# add new row
averages_row(df = df, cols = c(2, 4), label_col = 1, label = "Totals")
}
