% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-col.R
\name{format_}
\alias{format_}
\title{format_}
\usage{
format_(df, cols, fun, ...)
}
\arguments{
\item{df}{data frame containing the column to be formatted}

\item{cols}{numeric vector of column indices}

\item{fun}{formatting function}

\item{...}{additional arguments to pass to \code{fun}}
}
\description{
generalized function for formatting a data frame column
}

