% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-vec.R
\name{format_currency_vec}
\alias{format_currency_vec}
\title{format_currency_vec}
\usage{
format_currency_vec(x, digits = 2)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits afer the decimal point}
}
\description{
format numeric vectors with a currency symbol for reports
}
\examples{
x <- seq(1000, 3000, by = 1000)

format_currency_vec(x)
format_currency_vec(x, digits = 2)
}
