% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-col.R
\name{format_numeric}
\alias{format_numeric}
\title{format_numeric}
\usage{
format_numeric(df, cols, digits = 0)
}
\arguments{
\item{df}{a data frame}

\item{cols}{columns to format}

\item{digits}{number of digits after the decimal point}
}
\description{
format_numeric
}
\examples{

format_numeric(insurance_data, cols = 2, digits = 0)

}

