% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-vec.R
\name{format_numeric_vec}
\alias{format_numeric_vec}
\title{format_numeric_vec}
\usage{
format_numeric_vec(x, digits = 0)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits after the decimal point}
}
\description{
format numeric vectors for reports
}
\examples{
x <- seq(1000, 3000, by = 1000)

format_numeric_vec(x)
format_numeric_vec(x, digits = 2)
}
