% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-col.R
\name{format_origin}
\alias{format_origin}
\title{format_origin}
\usage{
format_origin(df, cols, eval, sep = " to ")
}
\arguments{
\item{df}{a data frame}

\item{cols}{columns to format}

\item{eval}{Most recent evaluation time. This will be used as the second
date in the last element in the vector \code{x}}

\item{sep}{character string to place between dates}
}
\description{
format_origin
}
\examples{
format_origin(insurance_data, cols = 1, eval = "2017")
}
