% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-vec.R
\name{format_origin_vec}
\alias{format_origin_vec}
\title{format_origin_vec}
\usage{
format_origin_vec(x, eval, sep = " to ")
}
\arguments{
\item{x}{a vector of dates.  Each date indicates the first day
of the origin period}

\item{eval}{Most recent evaluation time. This will be used as the second
date in the last element in the vector \code{x}}

\item{sep}{character string to place between dates}
}
\description{
format origin period vector
}
\examples{
x <- c("2014-01-01", "2015-01-01", "2016-01-01")

format_origin_vec(x, eval = "2015-03-31")
format_origin_vec(x, eval = "2015-03-31", sep = " - ")
}
