% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-col.R
\name{format_percent}
\alias{format_percent}
\title{format_percent}
\usage{
format_percent(df, cols, digits = 0)
}
\arguments{
\item{df}{a data frame}

\item{cols}{columns to format}

\item{digits}{number of digits after the decimal point}
}
\description{
format_percent
}
\examples{

}

