% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance_data}
\alias{insurance_data}
\title{Simple insurance sample data}
\format{A data frame with 3 ros and 4 variables
\describe{
  \item{year}{calendar year}
  \item{exposure}{exposure}
  \item{reported}{reported loss}
  \item{ultimate}{ultimate loss}
}}
\usage{
insurance_data
}
\description{
Simple insurance sample data
}
\keyword{datasets}
