% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-row.R
\name{summary_row}
\alias{summary_row}
\title{summary_row}
\usage{
summary_row(df, cols, fun, ..., rows = NA, label_col = NA,
  label = "Summary", new_row = TRUE, fill = NA)
}
\arguments{
\item{df}{the data frame}

\item{cols}{numeric vector of columns to include in the summary row}

\item{fun}{function to summarize data frame rows}

\item{...}{other arguments to be passed to the function specified in the
fun argument}

\item{rows}{numeric vector of rows to include in summary calculation}

\item{label_col}{column number of column to hold the summary row label
specified by the \code{label} argument}

\item{label}{summary row label}

\item{new_row}{boolean value indicating whether the summary row should
be a new row attached to the bottom of the data frame or if it should be 
combined with an already existing summary row.}

\item{fill}{value to put in columns not specified in the \code{cols} argument.
Default is NA}
}
\description{
add a summary row to a data frame
}
\examples{
df <- data.frame("year" = 2014:2016,
                 "a" = c(NA, 1, 3),
                 "b" = c("g", "e", "f")
      )

summary_row(fun = sum, na.rm = TRUE, df = df, cols = 2)
summary_row(fun = sum, na.rm = TRUE, df = df, cols = 2, label_col = 1, 
            label = "Totals")

df$d <- 5:7

# add new row
df <- summary_row(fun = sum, na.rm = TRUE, df = df, cols = c(2, 4), 
                  label_col = 1, label = "Totals")

# adjust newest row
df <- summary_row(fun = mean, na.rm = TRUE, df = df, cols = c(2, 4), 
                  label_col = 1, label = "Averages")

# I didn't want the mean in column 4, I actually wanted the product
df <- summary_row(fun = prod, na.rm = TRUE, df = df, cols = 4, label_col = 1, 
                  label = "Averages/Products", new_row = FALSE)
                  
# new cummary row with the average of the last 2 years
summary_row(fun = mean, df = df, cols = c(2, 4), rows = 2:3, label_col = 1, 
            label = "2 Yr AVG")
}

