% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-row.R
\name{averages_row}
\alias{averages_row}
\title{averages_row}
\usage{
averages_row(df, cols, rows = NA, label_col = NA, label = "Totals",
  new_row = TRUE, fill = NA)
}
\arguments{
\item{df}{the data frame}

\item{cols}{columns to include in the summary row}

\item{rows}{numeric vector of rows to include in summary calculation.  By default
\code{rows} is set to all rows that were not created by the \code{summary_row} 
function.}

\item{label_col}{column number of column to hold the summary row label
specified by the \code{label} argument}

\item{label}{summary row label}

\item{new_row}{boolean value indicating whether the summary row should
be a new row attached to the bottom of the data frame or if it should be 
combined with an already existing summary row.}

\item{fill}{value to put in columns not specified in the \code{cols} argument.
Default is NA}
}
\description{
add a averages summary row to a data frame
}
\examples{
df <- data.frame("year" = 2014:2016,
                 "a" = c(NA, 1, 3),
                 "b" = c("g", "e", "f")
      )

averages_row(df = df, cols = 2)
averages_row(df = df, cols = 2, label_col = 1, label = "Totals")

df$d <- 5:7

# add new row
averages_row(df = df, cols = c(2, 4), label_col = 1, label = "Totals")
}

