% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-vec.R
\name{format_percent_vec}
\alias{format_percent_vec}
\title{format_percent_vec}
\usage{
format_percent_vec(x, digits = 2)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits afer the decimal point}
}
\description{
format numeric vectors for reports
}
\examples{
x <- seq(1000, 3000, by = 1000)

format_percent_vec(x)
format_percent_vec(x, digits = 2)
}

